:truck: refactor: refactor package
