+++ .idea/workspace.xml
@@ -4,10 +4,12 @@
     <option name="autoReloadType" value="SELECTIVE" />
   </component>
   <component name="ChangeListManager">
-    <list default="true" id="77df237f-a12e-4eb1-bc1e-6cf2e42bf405" name="Changes" comment=":tada: init project">
-      <change afterPath="$PROJECT_DIR$/src/main/java/model/kiosk/Store.java" afterDir="false" />
+    <list default="true" id="77df237f-a12e-4eb1-bc1e-6cf2e42bf405" name="Changes" comment=":sparkles: feat: Repository interface">
       <change beforePath="$PROJECT_DIR$/.idea/workspace.xml" beforeDir="false" afterPath="$PROJECT_DIR$/.idea/workspace.xml" afterDir="false" />
-      <change beforePath="$PROJECT_DIR$/build.gradle" beforeDir="false" afterPath="$PROJECT_DIR$/build.gradle" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/src/main/java/controller/kiosk/KioskController.java" beforeDir="false" afterPath="$PROJECT_DIR$/src/main/java/kiosk/controller/kiosk/KioskController.java" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/src/main/java/controller/kiosk/KioskControllerImpl.java" beforeDir="false" afterPath="$PROJECT_DIR$/src/main/java/kiosk/controller/kiosk/KioskControllerImpl.java" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/src/main/java/model/kiosk/Store.java" beforeDir="false" afterPath="$PROJECT_DIR$/src/main/java/kiosk/model/store/Store.java" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/src/main/java/model/vo/Money.java" beforeDir="false" afterPath="$PROJECT_DIR$/src/main/java/common/vo/Money.java" afterDir="false" />
     </list>
     <option name="SHOW_DIALOG" value="false" />
     <option name="HIGHLIGHT_CONFLICTS" value="true" />
@@ -22,8 +24,8 @@
   <component name="FileTemplateManagerImpl">
     <option name="RECENT_TEMPLATES">
       <list>
-        <option value="Interface" />
         <option value="Class" />
+        <option value="Interface" />
       </list>
     </option>
   </component>
@@ -56,7 +58,7 @@
     "Gradle.KI5SK [dependencies].executor": "Run",
     "RunOnceActivity.ShowReadmeOnStart": "true",
     "dart.analysis.tool.window.visible": "false",
-    "git-widget-placeholder": "main",
+    "git-widget-placeholder": "feat/repository",
     "kotlin-language-version-configured": "true",
     "node.js.detected.package.eslint": "true",
     "node.js.detected.package.tslint": "true",
@@ -113,7 +115,7 @@
       <option name="number" value="Default" />
       <option name="presentableId" value="Default" />
       <updated>1712905876033</updated>
-      <workItem from="1712905877060" duration="18223000" />
+      <workItem from="1712905877060" duration="19933000" />
     </task>
     <task id="LOCAL-00001" summary=":tada: init project">
       <option name="closed" value="true" />
@@ -123,7 +125,23 @@
       <option name="project" value="LOCAL" />
       <updated>1712905908159</updated>
     </task>
-    <option name="localTasksCounter" value="2" />
+    <task id="LOCAL-00002" summary=":sparkles: feat: Money VO Object &amp; Skeleton">
+      <option name="closed" value="true" />
+      <created>1713143240429</created>
+      <option name="number" value="00002" />
+      <option name="presentableId" value="LOCAL-00002" />
+      <option name="project" value="LOCAL" />
+      <updated>1713143240429</updated>
+    </task>
+    <task id="LOCAL-00003" summary=":sparkles: feat: Repository interface">
+      <option name="closed" value="true" />
+      <created>1713144581720</created>
+      <option name="number" value="00003" />
+      <option name="presentableId" value="LOCAL-00003" />
+      <option name="project" value="LOCAL" />
+      <updated>1713144581720</updated>
+    </task>
+    <option name="localTasksCounter" value="4" />
     <servers />
   </component>
   <component name="TypeScriptGeneratedFilesManager">
@@ -131,6 +149,8 @@
   </component>
   <component name="VcsManagerConfiguration">
     <MESSAGE value=":tada: init project" />
-    <option name="LAST_COMMIT_MESSAGE" value=":tada: init project" />
+    <MESSAGE value=":sparkles: feat: Money VO Object &amp; Skeleton" />
+    <MESSAGE value=":sparkles: feat: Repository interface" />
+    <option name="LAST_COMMIT_MESSAGE" value=":sparkles: feat: Repository interface" />
   </component>
 </project>
