+++ src/main/java/kiosk/controller/kiosk/KioskControllerImpl.java
@@ -1,4 +1,4 @@
-package controller.kiosk;
+package kiosk.controller.kiosk;
 
 public class KioskControllerImpl {
 }
