+++ src/main/java/kiosk/model/store/Store.java
@@ -1,4 +1,4 @@
-package model.kiosk;
+package kiosk.model.store;
 
 public class Store {
     private String name;
